Refactor: background 까지 모듈화 분리 완료
