+++ Projects/App/Sources/DesignSystem/Form/TextField.swift
@@ -8,42 +8,42 @@
 
 import SwiftUI
 
-enum TFType {
-    /// 일반 텍스트 필드(회원가입 시 닉네임 입력받기)
-    case normal
-    /// 닉네임 입력시 발생하는 오류
-    case error
-}
-
-struct CustomTextFieldStyle: ViewModifier {
-    let type: TFType
-    func body(content: Content) -> some View {
-        switch type {
-        case .normal:
-            content
-                .padding(.vertical, 20)
-                .padding(.leading, 20)
-                .background(Color.symGray1)
-                .font(.medium(14))
-                .clipShape(RoundedRectangle(cornerRadius: 15))
-            
-        case .error:
-            content
-                .padding(.vertical, 20)
-                .padding(.leading, 20)
-                .font(.medium(14))
-                .background(Color.white)
-                .clipShape(RoundedRectangle(cornerRadius: 15))
-                .background(
-                    RoundedRectangle(cornerRadius: 15)
-                        .stroke(Color.errorRed, lineWidth: 1)
-                )
-        }
-    }
-}
-
-extension TextField {
-    func customTF(type: TFType) -> some View {
-        self.modifier(CustomTextFieldStyle(type: type))
-    }
-}
+//enum TFType {
+//    /// 일반 텍스트 필드(회원가입 시 닉네임 입력받기)
+//    case normal
+//    /// 닉네임 입력시 발생하는 오류
+//    case error
+//}
+//
+//struct CustomTextFieldStyle: ViewModifier {
+//    let type: TFType
+//    func body(content: Content) -> some View {
+//        switch type {
+//        case .normal:
+//            content
+//                .padding(.vertical, 20)
+//                .padding(.leading, 20)
+//                .background(Color.symGray1)
+//                .font(.medium(14))
+//                .clipShape(RoundedRectangle(cornerRadius: 15))
+//            
+//        case .error:
+//            content
+//                .padding(.vertical, 20)
+//                .padding(.leading, 20)
+//                .font(.medium(14))
+//                .background(Color.white)
+//                .clipShape(RoundedRectangle(cornerRadius: 15))
+//                .background(
+//                    RoundedRectangle(cornerRadius: 15)
+//                        .stroke(Color.errorRed, lineWidth: 1)
+//                )
+//        }
+//    }
+//}
+//
+//extension TextField {
+//    func customTF(type: TFType) -> some View {
+//        self.modifier(CustomTextFieldStyle(type: type))
+//    }
+//}
