+++ Projects/App/Sources/Presentation/Calendar/ViewModel/CalendarViewModel.swift
@@ -9,6 +9,7 @@
 import Foundation
 import CoreData
 import Combine
+import DesignSystem
 
 final class CalendarViewModel: RecordConditionFetch {
