+++ Projects/DesignSystem/Sources/DesignSystem/Color/Color.swift
@@ -11,23 +11,21 @@ import SwiftUI
 
 extension Color {
     // key color
-    static let main_FF7E7E = Color("Main")
-    static let sub_FFA9A9 = Color("Sub")
-    static let medium_FFF0F0 = Color("Medium")
-    static let bright_FFF3F3 = Color("Bright")
+    static let main_FF7E7E = Color(asset: DesignSystemAsset.Assets.main)
+    static let sub_FFA9A9 = Color(asset: DesignSystemAsset.Assets.sub)
+    static let medium_FFF0F0 = Color(asset: DesignSystemAsset.Assets.medium)
+    static let bright_FFF3F3 = Color(asset: DesignSystemAsset.Assets.bright)
     
     // gray scale
-    static let black2D2D2D = Color("Black")
-    static let gray6_42454A = Color("Gray6")
-    static let gray5_69707B = Color("Gray5")
-    static let gray4_9BA3AE = Color("Gray4")
-    static let gray3_CCD2DA = Color("Gray3")
-    static let gray2_E8EAED = Color("Gray2")
-    static let gray1_F3F5F8 = Color("Gray1")
+    static let black2D2D2D = Color(asset: DesignSystemAsset.Assets.black)
+    static let gray6_42454A = Color(asset: DesignSystemAsset.Assets.gray6)
+    static let gray5_69707B = Color(asset: DesignSystemAsset.Assets.gray5)
+    static let gray4_9BA3AE = Color(asset: DesignSystemAsset.Assets.gray4)
+    static let gray3_CCD2DA = Color(asset: DesignSystemAsset.Assets.gray3)
+    static let gray2_E8EAED = Color(asset: DesignSystemAsset.Assets.gray2)
+    static let gray1_F3F5F8 = Color(asset: DesignSystemAsset.Assets.gray1)
     
     // error color
-    static let buddhism_FF5959 = Color("Buddhism")
-    static let possibility_4CD964 = Color("Possibility")
-    
-    static let test = Color("Test")
+    static let buddhism_FF5959 = Color(asset: DesignSystemAsset.Assets.buddhism)
+    static let possibility_4CD964 = Color(asset: DesignSystemAsset.Assets.possilbility)
 }
